+++ src/lib/hmf_djerba/plugins/hmf/genomic_landscape/constants.py
@@ -9,13 +9,13 @@ TCGA_CODE = 'tcga_code'
 PURITY_INPUT = 'purity'
 DONOR = 'donor'
 MSI_FILE = 'msi_file'
-CTDNA_FILE = 'ctdna_file'
 SAMPLE_TYPE = 'sample_type'
 UNKNOWN_SAMPLE_TYPE = 'Unknown sample type'
 
 # Constants for render
 PURITY_REASON = 'purity'
 COVERAGE_REASON = 'coverage'
+CHORD_REASON = 'chord'
 
 # biomarker reportability
 CAN_REPORT_HRD = 'can_report_hrd'
@@ -23,12 +23,8 @@ CAN_REPORT_MSI = 'can_report_msi'
 CANT_REPORT_HRD_REASON = 'cant_report_hrd_reason'
 
 # For MSI file
-MSI_RESULTS_SUFFIX = '.recalibrated.msi.booted'
-MSI_WORKFLOW = 'msisensor'
-
-# For ctDNA file
-CTDNA_RESULTS_SUFFIX = 'SNP.count.txt'
-CTDNA_WORKFLOW = 'mrdetect_filter_only'
+MSI_RESULTS_SUFFIX = '.purple.purity.tsv'
+MSI_WORKFLOW = 'purple'
 
 HRD_WORKFLOW = 'hrDetect'
 
@@ -43,10 +39,6 @@ CANCER_SPECIFIC_COHORT = 'Cancer-specific Cohort'
 CANCER_TYPE_HEADER = 'CANCER.TYPE'
 CLINICALLY_RELEVANT_VARIANTS = 'Clinically relevant variants'
 COMPASS = 'COMPASS'
-CTDNA = 'ctDNA'
-CTDNA_CANDIDATES = 'ctDNA_candidate_sites'
-CTDNA_ELIGIBILITY = 'ctDNA_eligibility'
-CTDNA_ELIGIBILITY_CUTOFF = 4000
 DATA_SEGMENTS = 'data.seg'
 GENOME_SIZE = 3*10**9 # TODO use more accurate value when we release a new report format
 GENOMIC_BIOMARKERS = 'genomic_biomarkers.maf'
@@ -61,9 +53,7 @@ METRIC_VALUE = 'Genomic biomarker value'
 MINIMUM_MAGNITUDE_SEG_MEAN = 0.2
 MRDETECT_FILTER_ONLY_FILE_NAME = 'SNP.count.txt'
 MSI = "MSI"
-MSI_CUTOFF = 15.0
 MSI_FILE_NAME = 'msi.txt'
-MSS_CUTOFF = 5.0
 MUTATIONS_EXTENDED = 'data_mutations_extended.txt'
 NA = 'NA'
 ONCOKB = 'OncoKB'
@@ -96,4 +86,7 @@ HRD = 'HRD'
 HRD_short = 'HRD_short'
 HRD_long = 'HRD_long'
 HRD_PLOT = 'hrd_base64'
-HRDETECT_PATH = 'hrd_path'
+CHORD_PATH = 'hrd_path'
+HR_DEFICIENT = 'HR_deficient'
+HR_PROFICIENT = 'HR_proficient'
+HRD_PLOT_FILENAME = 'HRD_plot.png'