+++ src/lib/hmf_djerba/plugins/hmf/wgts/cnv_purple/legacy_tools.py
@@ -7,12 +7,13 @@ import json
 import logging
 import os
 import djerba.core.constants as core_constants
-import djerba.plugins.wgts.cnv_purple.legacy_constants as cnv
+import hmf_djerba.plugins.hmf.wgts.cnv_purple.legacy_constants as cnv
 import djerba.util.oncokb.constants as oncokb_constants
 from djerba.mergers.gene_information_merger.factory import factory as gim_factory
 from djerba.mergers.treatment_options_merger.factory import factory as tom_factory
 from djerba.util.sequenza import sequenza_reader 
 from djerba.util.environment import directory_finder
+from djerba.util.expression_reader import expression_reader
 from djerba.util.html import html_builder
 from djerba.util.image_to_base64 import converter
 from djerba.util.logger import logger
@@ -20,13 +21,14 @@ from djerba.util.oncokb.annotator import annotator_factory
 from djerba.util.oncokb.tools import levels as oncokb_levels
 from djerba.util.oncokb.tools import gene_summary_reader
 from djerba.util.subprocess_runner import subprocess_runner
-from djerba.util.wgts.tools import wgts_tools
+from djerba.util.variant_sorter import variant_sorter
+
 
 class cnv_processor(logger):
 
     ALTERATION_UPPER_CASE = 'ALTERATION'
     HUGO_SYMBOL_UPPER_CASE = 'HUGO_SYMBOL'
-    PLOT_FILENAME = 'seg_CNV_plot.svg'
+    PLOT_FILENAME = 'seg_CNV_plot.png'
     MINIMUM_MAGNITUDE_SEG_MEAN = 0.2
     GENOME_SIZE = 3095978931 # comes from https://www.ncbi.nlm.nih.gov/grc/human/data?asm=GRCh38.p12. Non-N bases. 
     SEG_FILENAME = 'seg.txt'
@@ -106,35 +108,41 @@ class cnv_processor(logger):
     def get_results(self):
         """Read previous output into the JSON serializable results structure"""
         image_converter = converter(self.log_level, self.log_path)
-        cnv_plot = image_converter.convert_svg(self.plot_path, 'CNV plot')
+        cnv_plot = image_converter.convert_png(self.plot_path, 'CNV plot')
         rows = []
-        wgts_toolkit = wgts_tools(self.log_level, self.log_path)
-        is_wgts = wgts_toolkit.has_expression(self.work_dir)
+        xreader = expression_reader(self.log_level, self.log_path)
+        is_wgts = xreader.has_expression(self.work_dir)
         if is_wgts:
             self.logger.info("Reading expression from {0}".format(self.work_dir))
-            mutation_expression = wgts_tools.read_expression(self.work_dir)
+            mutation_expression = xreader.read_expression(self.work_dir)
         else:
             self.logger.info("No expression data found")
             mutation_expression = {}
-        cytobands = wgts_tools(self.log_level, self.log_path).cytoband_lookup()
+        var_sorter = variant_sorter(self.log_level, self.log_path)
+        cytobands = var_sorter.cytoband_lookup()
         input_name = oncokb_constants.DATA_CNA_ONCOKB_GENES_NON_DIPLOID_ANNOTATED
         with open(os.path.join(self.work_dir, input_name)) as input_file:
             reader = csv.DictReader(input_file, delimiter="\t")
+            seen_entries = set()
             for row_input in reader:
                 gene = row_input[self.HUGO_SYMBOL_UPPER_CASE]
-                # if gene not found in cytoBands.txt, default to 'Unknown'
-                row_output = {
-                    cnv.EXPRESSION_PERCENTILE: mutation_expression.get(gene), # None for WGS
-                    wgts_tools.GENE: gene,
-                    cnv.GENE_URL: html_builder.build_gene_url(gene),
-                    cnv.ALTERATION: row_input[self.ALTERATION_UPPER_CASE],
-                    wgts_tools.CHROMOSOME: cytobands.get(gene, wgts_tools.UNKNOWN),
-                    wgts_tools.ONCOKB: oncokb_levels.parse_oncokb_level(row_input)
-                }
-                rows.append(row_output)
+                alteration = row_input[self.ALTERATION_UPPER_CASE]
+                entry = (gene, alteration)
+                if entry not in seen_entries:
+                    seen_entries.add(entry)
+                    # if gene not found in cytoBands.txt, default to 'Unknown'
+                    row_output = {
+                        cnv.EXPRESSION_PERCENTILE: mutation_expression.get(gene), # None for WGS
+                        var_sorter.GENE: gene,
+                        cnv.GENE_URL: html_builder.build_gene_url(gene),
+                        cnv.ALTERATION: alteration,
+                        var_sorter.CHROMOSOME: cytobands.get(gene, var_sorter.UNKNOWN),
+                        var_sorter.ONCOKB: oncokb_levels.parse_oncokb_level(row_input)
+                    }
+                    rows.append(row_output)
         unfiltered_cnv_total = len(rows)
         self.logger.debug("Sorting and filtering CNV rows")
-        rows = wgts_toolkit.sort_variant_rows(rows)
+        rows = var_sorter.sort_variant_rows(rows)
         rows = oncokb_levels.filter_reportable(rows)
         results = {
             cnv.PERCENT_GENOME_ALTERED: self.calculate_percent_genome_altered(),